move osrm_api to separate module
