+++ services/travelmux/src/api/v5/mod.rs
@@ -1,6 +1,8 @@
 mod error;
+mod osrm_api;
 pub mod plan;
 mod travel_modes;
+
 pub use travel_modes::TravelModes;
 
 pub use plan::{Itinerary, Plan};
